web-fragment.xml removed from openxava, replaced by a Spring auto-configuration
