+++ openxava/pom.xml
@@ -279,6 +279,18 @@
 			<artifactId>commonmark</artifactId>
 			<version>0.24.0</version>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot</artifactId>
+			<version>4.1.0</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-autoconfigure</artifactId>
+			<version>4.1.0</version>
+			<scope>provided</scope>
+		</dependency>
 
 	</dependencies>
