--- openxava/src/main/resources/META-INF/web-fragment.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-
-<!--
-NaviOX. Navigation and security for OpenXava applications.
-Copyright 2014 Javier Paniza Lucas
-
-License: Apache License, Version 2.0.	
-You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-License: GNU Lesser General Public License (LGPL), version 2.1 or later.
-See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
--->
-
-<web-fragment xmlns="https://jakarta.ee/xml/ns/jakartaee"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee
-    https://jakarta.ee/xml/ns/jakartaee/web-fragment_6_0.xsd"
-    version="6.0">
-	
-	<filter>
-		<filter-name>naviox</filter-name>
-		<filter-class>com.openxava.naviox.web.NaviOXFilter</filter-class>
-	</filter>
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<url-pattern>*.jsp</url-pattern>
-	</filter-mapping>		
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<url-pattern>/modules/*</url-pattern>
-		<dispatcher>REQUEST</dispatcher>
-		<dispatcher>FORWARD</dispatcher>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<url-pattern>/phone/index.jsp</url-pattern>
-		<dispatcher>REQUEST</dispatcher>
-		<dispatcher>FORWARD</dispatcher>
-	</filter-mapping>		
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<servlet-name>naviox</servlet-name>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<servlet-name>module</servlet-name>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>naviox</filter-name>
-		<servlet-name>phone</servlet-name>
-	</filter-mapping>
-			
-</web-fragment>
